Improve UPX detection by section names.
To prevent false positives, we should check whether the size of the very
first section of the file is zero. If it is not, then we either have a
modified packer, already unpacked file that has the packer's section
names, or a file that tries to pretend it is packed with UPX. This should
reduce the amount of false positives significantly.